www/requirements.txt
fastapi
uvicorn
python-multipart
matplotlib
numpy
Pillow
